+++ assets/js/components/Vehicles/Options.vue
@@ -23,13 +23,16 @@
 	</label>
 </template>
 
-<script>
-export default {
+<script lang="ts">
+import type { Vehicle } from "@/types/evcc";
+import { defineComponent, type PropType } from "vue";
+
+export default defineComponent({
 	name: "VehicleOptions",
 	props: {
 		connected: Boolean,
 		id: [String, Number],
-		vehicles: Array,
+		vehicles: Array as PropType<Vehicle[]>,
 		selected: String,
 	},
 	emits: ["change-vehicle", "remove-vehicle"],
@@ -39,8 +42,8 @@ export default {
 		},
 	},
 	methods: {
-		change(event) {
-			const name = event.target.value;
+		change(event: Event) {
+			const name = (event.target as HTMLSelectElement).value;
 			if (name) {
 				this.$emit("change-vehicle", name);
 			} else {
@@ -48,7 +51,7 @@ export default {
 			}
 		},
 	},
-};
+});
 </script>
 <style scoped>
 .custom-select {
