+++ assets/js/components/Vehicles/Title.vue
@@ -51,7 +51,7 @@
 	</div>
 </template>
 
-<script>
+<script lang="ts">
 import "@h2d2/shopicons/es/regular/cablecharge";
 import Tooltip from "bootstrap/js/dist/tooltip";
 import Modal from "bootstrap/js/dist/modal";
@@ -60,8 +60,10 @@ import Options from "./Options.vue";
 import CloudOffline from "../MaterialIcon/CloudOffline.vue";
 import Sync from "../MaterialIcon/Sync.vue";
 import collector from "@/mixins/collector";
+import { defineComponent, type PropType } from "vue";
+import type { Vehicle } from "@/types/evcc";
 
-export default {
+export default defineComponent({
 	name: "VehicleTitle",
 	components: { VehicleOptions: Options, VehicleIcon, Sync, CloudOffline },
 	mixins: [collector],
@@ -72,14 +74,14 @@ export default {
 		vehicleNotReachable: Boolean,
 		icon: String,
 		vehicleName: String,
-		vehicles: { type: Array, default: () => [] },
+		vehicles: { type: Array as PropType<Vehicle[]>, default: () => [] },
 		title: String,
 	},
 	emits: ["change-vehicle", "remove-vehicle"],
 	data() {
 		return {
-			refreshTooltip: null,
-			notReachableTooltip: null,
+			refreshTooltip: null as Tooltip | null,
+			notReachableTooltip: null as Tooltip | null,
 		};
 	},
 	computed: {
@@ -120,7 +122,7 @@ export default {
 		this.initTooltip();
 	},
 	methods: {
-		changeVehicle(name) {
+		changeVehicle(name: string) {
 			this.$emit("change-vehicle", name);
 		},
 		removeVehicle() {
@@ -130,21 +132,23 @@ export default {
 			this.$nextTick(() => {
 				this.refreshTooltip?.dispose();
 				this.notReachableTooltip?.dispose();
-				if (this.$refs.refresh) {
-					this.refreshTooltip = new Tooltip(this.$refs.refresh);
+				if (this.$refs["refresh"]) {
+					this.refreshTooltip = new Tooltip(this.$refs["refresh"]);
 				}
-				if (this.$refs.notReachable) {
-					this.notReachableTooltip = new Tooltip(this.$refs.notReachable);
+				if (this.$refs["notReachable"]) {
+					this.notReachableTooltip = new Tooltip(this.$refs["notReachable"]);
 				}
 			});
 		},
 		openHelpModal() {
-			const modal = Modal.getOrCreateInstance(document.getElementById("helpModal"));
+			const modal = Modal.getOrCreateInstance(
+				document.getElementById("helpModal") as HTMLElement
+			);
 			modal.show();
 			this.initTooltip();
 		},
 	},
-};
+});
 </script>
 
 <style scoped>
